rasa_python/rasa_v2/requirements.txt
absl-py==0.10.0
aiofiles==0.5.0
aiohttp==3.6.3
APScheduler==3.6.3
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
blis==0.4.1
boto3==1.15.18
botocore==1.18.18
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.3
characteristic==14.3.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.4
colorclass==2.2.0
coloredlogs==14.0
colorhash==1.0.2
constantly==15.1.0
cryptography==3.1.1
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
dm-tree==0.1.5
dnspython==1.16.0
docker==4.3.1
docopt==0.6.2
english==2020.7.0
entrypoints==0.3
fbmessenger==6.0.0
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gast==0.3.3
gevent==21.1.2
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
h11==0.8.1
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
httpcore==0.3.0
httplib2==0.18.1
httptools==0.1.1
httpx==0.15.4
humanfriendly==8.2
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==2.0.0
incremental==21.3.0
ipykernel==5.3.4
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kafka-python==2.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
klein==20.6.0
Markdown==3.3.1
markdownify==0.5.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mattermostwrapper==2.2
mistune==0.8.4
mitie==0.7.0
monotonic==1.5
multidict==4.7.6
murmurhash==1.0.2
mysqlclient==2.0.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
neotime==1.7.4
nest-asyncio==1.4.1
networkx==2.5
notebook==6.1.4
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandocfilters==1.4.2
pansi==2020.7.3
parso==0.7.0
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
Pillow==8.0.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==2.0.10
protobuf==3.13.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py2neo==2020.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
Pygments==2.7.1
PyJWT==1.7.1
pykwalify==1.7.0
pymongo==3.10.1
PyMySQL==1.0.2
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.17.3
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
python-telegram-bot==12.8
pytz==2020.1
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
questionary==1.4.0
rasa==2.0.2
rasa-nlu==0.13.3
rasa-sdk==2.5.0
redis==3.5.3
regex==2020.9.27
requests==2.24.0
requests-async==0.5.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rocketchat-API==1.9.1
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
sacremoses==0.0.43
sanic==19.9.0
Sanic-Cors==0.10.0.post3
sanic-jwt==1.4.1
Sanic-Plugins-Framework==0.9.4
scikit-learn==0.23.2
scipy==1.5.3
Send2Trash==1.5.0
sentencepiece==0.1.94
sentry-sdk==0.17.8
simplejson==3.17.2
six==1.15.0
sklearn-crfsuite==0.3.6
slackclient==2.9.2
sniffio==1.2.0
soupsieve==2.0.1
spacy==2.2.4
SQLAlchemy==1.3.20
srsly==1.0.2
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0
tensorflow-hub==0.9.0
tensorflow-probability==0.11.1
tensorflow-text==2.3.0
termcolor==1.1.0
terminado==0.9.1
terminaltables==3.1.0
testpath==0.4.4
thinc==7.4.0
threadpoolctl==2.1.0
tokenizers==0.9.2
tornado==6.0.4
tqdm==4.50.2
transformers==3.4.0
Tubes==0.2.0
twilio==6.45.4
Twisted==21.2.0
typeguard==2.9.1
typing==3.7.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==3.2.0
urllib3==1.25.10
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
webexteamssdk==1.6
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
yarl==1.5.1
zipp==3.3.1
zope.event==4.5.0
zope.interface==5.4.0
